requirements/test.txt
-r base.txt

pytest==3.4.1
pytest-travis-fold==1.3.0
pytest-cov==2.5.1
pytest-sugar==0.9.1
pytest-django==3.1.2
django-fake-model==0.1.4

